Introduced cmake for building on linux
The big change here is that I have moved the headers into what was called the
src/ directory and renamed src/ to modus/. I also added CMakeLists.txt files to
the top level directory and to modus/. To build the code on linux, go to the top
level directory,

$ mkdir build
$ cd build
$ cmake -DMAKE_INSTALL_PREFIX=install ..
$ make all install

This will create an out-of-source build directory and build the modus library
under build/root/lib. The install prefix defines an installation directory
underneath the build/ directory, which is a nice safe place to have it during
development. When you want to install modus so that the rest of your system can
use it, change the prefix to /usr/local.